third_party/blink/web_tests/fast/events/touch/gesture/gesture-tap-click-common-ancestor-expected.txt

Verifies that each mouse event that occurs when tapping on an element reflects the node under the point at the time of that event, and the click event goes to the common ancestor of the mousedown and mouseup events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Sending GestureTapDown
Sending GestureShowPress
Sending GestureTap
Received mousemove on child3
PASS event.target.id is "child3"
Received mousedown on child2
PASS event.target.id is "child2"
Received mouseup on child1
PASS event.target.id is "child1"
Received click on ancestor12
PASS event.target.id is "ancestor12"
PASS successfullyParsed is true

TEST COMPLETE
